backend/pyproject.toml
[project]
name = "beacon-backend"
version = "0.1.0"
description = "990 Beacon API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.7.0",
    "svix>=1.60.0",
    "httpx>=0.28.0",
    "psycopg[binary]>=3.2.0",
    "fastapi-clerk-auth>=0.0.9",
    "lxml>=5.0.0",
    "defusedxml>=0.7.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_test_loop_scope = "session"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 88
exclude = ["alembic/"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["B008"]

[tool.ruff.lint.per-file-ignores]
"tests/ingest/test_xml_parser.py" = ["N801"]

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
